More updates - it works even more now! (most of the game works, just some last-minute fixes to do)

+++ Button.cs
@@ -12,38 +12,38 @@ namespace FishGame
 {
     class Button
     {
-        // Representerar en knapp. Då MonoGame inte har en egen knapphanterare så har jag skapat en egen.
+        // Representerar en knapp. Då MonoGame inte har en egen knapphanterare så har jag skapat en egen!
 
         public Tuple<Texture2D, Texture2D> AssociatedAssets { get; set; } // Bilder som associeras med knappen när man har muspekaren över den respektive när man inte har det
         public Vector2 Position { get; set; } // Knappens position på skärmen
 
         public GameState ActiveOnScreen { get; set; } // Vilken skärm (titelskärmen, spelskärmen, etc.) som knappen finns tillgänglig på
         public Texture2D ActiveAsset { get; set; } // Den aktiva texturen som knappen visar
-        public int Scale { get; set; } // Vilken skala knappen ritas ut i
+        public float Scale { get; set; } // Vilken skala knappen ritas ut i
 
         private SoundEffectInstance _buttonClickSound { get; set; } // Ljud när knapp klickas på
         private SoundEffectInstance _buttonHoverSound { get; set; } // Ljud när knapp hålls över
 
-        private enum ButtonStates // Knappens möjliga statusar
+        private enum ButtonStates // Knappens möjliga statusar (relaterat till muspekaren)
         {
             Idle,
             Hovered,
             Clicked
         }
 
-        private ButtonStates _previousButtonState { get; set; }
+        private ButtonStates _previousButtonState { get; set; } // Knappens tidigare status
 
         public Button(
             string baseAssociatedAssetName,
             ContentManager contentLoader,
             Vector2 position,
             GameState activeOnScreen,
-            int scale = 1
+            float scale = 1
         )
         {
             AssociatedAssets = new Tuple<Texture2D, Texture2D>(
                 contentLoader.Load<Texture2D>(baseAssociatedAssetName),
-                contentLoader.Load<Texture2D>(baseAssociatedAssetName + "-hovered")
+                contentLoader.Load<Texture2D>(baseAssociatedAssetName + "-hovered") // Knappen ska ha en -hovered-bild som visas när man håller muspekaren över knappen
             ); // Hämta bilder till knappen
             ActiveAsset = AssociatedAssets.Item1;
             Scale = scale;
@@ -85,7 +85,7 @@ namespace FishGame
                     if (
                         _buttonHoverSound.State != SoundState.Playing
                         && _previousButtonState != ButtonStates.Hovered
-                    )
+                    ) // Om knappen hålls över, spela ett ljud som indikerar att knappen hålls över
                     {
                         _buttonHoverSound.Play();
                     }
@@ -122,7 +122,8 @@ namespace FishGame
         /// Funktion för att kontrollera om en knapp har klickats på eller inte.
         /// </summary>
         /// <param name="mouseState">Objekt för att hämta musens koordinater</param>
-        /// <param name="previousMouseState">Objekt som representerar musens tidigare status. Används för att motverka att knappen markeras som klickad när muspekaren tidigare hållts nere utan att släppas</param>
+        /// <param name="previousMouseState">Objekt som representerar musens tidigare status. Används för att motverka att
+        /// knappen markeras som klickad när muspekaren tidigare hållts nere utan att släppas när den dragits över knappen.</param>
         /// <returns></returns>
         public bool IsClicked(MouseState mouseState, MouseState previousMouseState)
         {
